functional-tests/cluster-1m2s/requirements.txt
dnspython==2.3.0
pytest==7.2.1
pytest-order==1.0.1
PyYAML==6.0
requests==2.28.2
